% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_sem.R
\name{fit_multiple_sem}
\alias{fit_multiple_sem}
\title{Title}
\usage{
fit_multiple_sem(models, data, ...)
}
\arguments{
\item{models}{A vector with multiple models}

\item{data}{data argument in sem}

\item{...}{other argument in sem}
}
\value{
summaries as in sem()
}
\description{
Title
}
\examples{
model1 <- 'A =~ x1 + x2 + x3'
model2 <- 'B =~ x4 + x5 + x6'
model3 <- 'C =~ x7 + x8 + x9'
models <- c(model1, model2, model3)
data("HolzingerSwineford1939", package = "lavaan")
summaries <- fit_multiple_sem(models, HolzingerSwineford1939)
}
