% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_cfa.R
\name{fit_multiple_cfa}
\alias{fit_multiple_cfa}
\title{Title}
\usage{
fit_multiple_cfa(models, data, ...)
}
\arguments{
\item{models}{A vector with multiple models}

\item{data}{data argument in cfa}

\item{...}{other argument in cfa}
}
\value{
results as in cfa()
}
\description{
Title
}
\examples{
models <- c(model1, model2, model3)
summaries <- fit_multiple_cfa(models, HolzingerSwineford1939)
}
