% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_sem.R
\name{fit_multiple_sem}
\alias{fit_multiple_sem}
\title{Title}
\usage{
fit_multiple_sem(models, data, ...)
}
\arguments{
\item{models}{A vector with multiple models}

\item{data}{data argument in sem}

\item{...}{other argument in sem}
}
\value{
results as in sem()
}
\description{
Title
}
\examples{
models <- c(model1, model2, model3)
summaries <- fit_multiple_sem(models, HolzingerSwineford1939)
}
